api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
asyncpg==0.29.0
pandas==2.1.3
numpy==1.26.2
aiohttp==3.9.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
openpyxl==3.1.2
pyarrow==14.0.1
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
pydantic-settings==2.0.3
python-jose==3.3.0
python-dotenv==1.0.0
python-multipart==0.0.6
bcrypt==4.0.1
pandas==2.1.0
numpy==1.26.0
scikit-learn==1.3.1
pgvector==0.2.0
openai==1.5.0
psycopg2-binary==2.9.9
httpx==0.25.0
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
ydata-profiling==4.6.0
timescaledb-python==0.3.0
matplotlib==3.7.2
seaborn==0.12.2
fastapi-pagination==0.12.10
prometheus-client==0.17.1
fastapi-cache2==0.2.1
redis==5.0.0
fastapi-limiter==0.1.5
faiss-cpu==1.7.4
python-dateutil==2.8.2
great-expectations==0.17.18
dask==2023.9.0
polars==0.18.2
fastapi-prometheus-exporter==2.0.0
